Method overriding 2.java
<<<Sample output is here>>>
A real example of Java Method Overriding
Consider a scenario where Bank is a class that provides functionality to get the rate of interest. However, the rate of interest varies according to banks. For example, SBI, ICICI and AXIS banks could provide 8%, 7%, and 9% rate of interest.
_________________________________
SAMPLE OUTPUT:
SBI Rate of Interest: 8ICICI Rate of Interest: 7AXIS Rate of Interest: 9